Two small changes to allow compile on SGI.  Adding an inline keyword and changing a size_t to unsigned.
Changes pass Unit Regression on Sun beta.  SGI .95 compiles, but untested on
SGI.

+++ src/Mesh/CornerTag.cpp
@@ -74,7 +74,7 @@ TagHandle CornerTagHandles::get_handle( Mesh* mesh, unsigned corners, MsqError& 
     if (cornerHandles[corners])
     {
       Mesh::TagType type;
-      size_t size;
+      unsigned size;
       mesh->tag_properties( cornerHandles[corners], name, type, size, err );
       MSQ_ERRZERO(err);
